requirements.txt
fastapi
uvicorn
openai
python-dotenv
twilio
requests
